fixed type issue

+++ python-sims/binarySearch1.py
@@ -14,7 +14,8 @@ import numpy as np
 
 N_values = np.zeros(11)
 for i in range(np.size(N_values)):
-    N_values[i] = 10000 * np.power(2, i)
+    N_values[i] = (10000 * int(np.power(2, i)))
+N_values = N_values.astype(np.int64)
 print("hello")
 estThresholdDegrees, estThresholdVs = ER_BinSearchThreshold_v(32, N_values)
 print(estThresholdDegrees)
